Create PlotlyJSONGenerator ReportGenerator
This is a ReportGenerator which creates JSON files for [plotly.js] a
declarative charting library built on d3.js and stack.gl.  It can be
used by skins to create interactive graphs/charts.

It is based on ImageGenerator and shares a significant amount of code.
These classes could be refactored to add a common base class to remove
much of this duplication.

It uses the ImageGenerator configuration from skin.conf, since it serves
the same purpose and supports nearly all of the properties.

[plotly.js]: https://plot.ly/javascript/

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>